+++ components/form-builder/app/edit/index.ts
@@ -1,4 +1,4 @@
-export { Elements } from "./Elements";
+export { Elements } from "./Edit";
 export { ElementPanel } from "./ElementPanel";
 export { PanelActions } from "./PanelActions";
 export { PanelActionsLocked } from "./PanelActionsLocked";
